tests: fold byte-cap regression tests into each source's test file
Moves the per-source byte-cap tests out of standalone bytecap_test.go files
into each package's conventional <source>_test.go (merging where one already
exists, renaming otherwise). No behavior change; tests are unchanged.
