--- pkg/source/espn/bytecap_test.go
@@ -1,63 +1,0 @@
-package espn
-
-import (
-	"context"
-	"encoding/json"
-	"net/http"
-	"net/http/httptest"
-	"strings"
-	"testing"
-
-	httpclient "github.com/bruin-data/ingestr/pkg/http"
-	"github.com/bruin-data/ingestr/pkg/source"
-)
-
-func TestESPNByteCap(t *testing.T) {
-	wide := strings.Repeat("x", 2048)
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		articles := []map[string]interface{}{}
-		for i := 0; i < 50; i++ {
-			articles = append(articles, map[string]interface{}{"id": i, "name": wide})
-		}
-		w.Header().Set("Content-Type", "application/json")
-		_ = json.NewEncoder(w).Encode(map[string]interface{}{"articles": articles})
-	}))
-	defer srv.Close()
-
-	run := func(max int64) (int64, int64) {
-		s := &ESPNSource{
-			client:  httpclient.New(httpclient.WithBaseURL(srv.URL)),
-			sport:   "football",
-			league:  "nfl",
-			limit:   100,
-			baseURL: srv.URL,
-		}
-		cfg := s.tables()["news"]
-		results, err := s.read(context.Background(), cfg, source.ReadOptions{MaxBatchBytes: max})
-		if err != nil {
-			t.Fatal(err)
-		}
-		var b, rw int64
-		for res := range results {
-			if res.Err != nil {
-				t.Fatal(res.Err)
-			}
-			b++
-			rw += res.Batch.NumRows()
-			res.Batch.Release()
-		}
-		return b, rw
-	}
-
-	offB, offR := run(0)
-	onB, onR := run(4096)
-	if offB != 1 {
-		t.Fatalf("cap-off batches=%d want 1", offB)
-	}
-	if onB <= 1 {
-		t.Fatalf("cap-on batches=%d want >1", onB)
-	}
-	if offR != onR || offR != 50 {
-		t.Fatalf("row mismatch off=%d on=%d", offR, onR)
-	}
-}
