+++ pkg/source/frankfurter/frankfurter_test.go
@@ -3,8 +3,11 @@ package frankfurter
 import (
 	"bytes"
 	"context"
+	"encoding/json"
+	"fmt"
 	"net/http"
 	"net/http/httptest"
+	"strings"
 	"testing"
 	"time"
 
@@ -241,5 +244,48 @@ func TestToDateString(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			assert.Equal(t, tt.want, toDateString(tt.input))
 		})
+	}
+}
+
+func TestFrankfurterByteCap(t *testing.T) {
+	wide := strings.Repeat("x", 2048)
+	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		currencies := map[string]string{}
+		for i := 0; i < 50; i++ {
+			currencies[fmt.Sprintf("C%02d", i)] = wide
+		}
+		w.Header().Set("Content-Type", "application/json")
+		_ = json.NewEncoder(w).Encode(currencies)
+	}))
+	defer srv.Close()
+
+	run := func(max int64) (int64, int64) {
+		s := &FrankfurterSource{client: ingestrhttp.New(ingestrhttp.WithBaseURL(srv.URL))}
+		results, err := s.read(context.Background(), "currencies", "EUR", source.ReadOptions{MaxBatchBytes: max})
+		if err != nil {
+			t.Fatal(err)
+		}
+		var b, rw int64
+		for res := range results {
+			if res.Err != nil {
+				t.Fatal(res.Err)
+			}
+			b++
+			rw += res.Batch.NumRows()
+			res.Batch.Release()
+		}
+		return b, rw
+	}
+
+	offB, offR := run(0)
+	onB, onR := run(4096)
+	if offB != 1 {
+		t.Fatalf("cap-off batches=%d want 1", offB)
+	}
+	if onB <= 1 {
+		t.Fatalf("cap-on batches=%d want >1", onB)
+	}
+	if offR != onR || offR != 50 {
+		t.Fatalf("row mismatch off=%d on=%d", offR, onR)
 	}
 }
