+++ pkg/source/redditads/redditads_test.go
@@ -3,9 +3,14 @@ package redditads
 import (
 	"context"
 	"encoding/json"
+	"net/http"
+	"net/http/httptest"
 	"strings"
 	"testing"
 	"time"
+
+	httpclient "github.com/bruin-data/ingestr/pkg/http"
+	"github.com/bruin-data/ingestr/pkg/source"
 )
 
 func TestFilterItemsByInterval(t *testing.T) {
@@ -572,5 +577,53 @@ func TestSplitAndTrim(t *testing.T) {
 				t.Fatalf("splitAndTrim(%q)[%d]: expected %q, got %q", tt.input, i, want, got[i])
 			}
 		}
+	}
+}
+
+func TestRedditAdsByteCap(t *testing.T) {
+	wide := strings.Repeat("x", 2048)
+	calls := 0
+	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		calls++
+		rows := []map[string]interface{}{}
+		if calls == 1 {
+			for i := 0; i < 50; i++ {
+				rows = append(rows, map[string]interface{}{"id": i, "name": wide})
+			}
+		}
+		w.Header().Set("Content-Type", "application/json")
+		_ = json.NewEncoder(w).Encode(map[string]interface{}{"data": rows})
+	}))
+	defer srv.Close()
+
+	run := func(max int64) (int64, int64) {
+		s := &RedditAdsSource{client: httpclient.New(httpclient.WithBaseURL(srv.URL))}
+		calls = 0
+		results, err := s.read(context.Background(), "campaigns", []string{"acct1"}, source.ReadOptions{MaxBatchBytes: max})
+		if err != nil {
+			t.Fatal(err)
+		}
+		var b, rw int64
+		for res := range results {
+			if res.Err != nil {
+				t.Fatal(res.Err)
+			}
+			b++
+			rw += res.Batch.NumRows()
+			res.Batch.Release()
+		}
+		return b, rw
+	}
+
+	offB, offR := run(0)
+	onB, onR := run(4096)
+	if offB != 1 {
+		t.Fatalf("cap-off batches=%d want 1", offB)
+	}
+	if onB <= 1 {
+		t.Fatalf("cap-on batches=%d want >1", onB)
+	}
+	if offR != onR || offR != 50 {
+		t.Fatalf("row mismatch off=%d on=%d", offR, onR)
 	}
 }
