+++ pkg/source/smartsheet/smartsheet_test.go
@@ -2,9 +2,13 @@ package smartsheet
 
 import (
 	"context"
+	"encoding/json"
+	"net/http"
+	"net/http/httptest"
 	"strings"
 	"testing"
 
+	httpclient "github.com/bruin-data/ingestr/pkg/http"
 	"github.com/bruin-data/ingestr/pkg/source"
 )
 
@@ -154,5 +158,56 @@ func TestGetTable_PassesResolvedSheetID(t *testing.T) {
 	}
 	if tbl.Name() != "111" {
 		t.Errorf("expected sheet ID 111 (from URI via 'sheet' alias), got %s", tbl.Name())
+	}
+}
+
+func TestSmartsheetByteCap(t *testing.T) {
+	wide := strings.Repeat("x", 2048)
+	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		rows := make([]map[string]interface{}, 0, 50)
+		for i := 0; i < 50; i++ {
+			rows = append(rows, map[string]interface{}{
+				"id":    i,
+				"cells": []map[string]interface{}{{"columnId": 1, "value": wide}},
+			})
+		}
+		body := map[string]interface{}{
+			"name":    "sheet",
+			"columns": []map[string]interface{}{{"id": 1, "title": "name", "type": "TEXT_NUMBER"}},
+			"rows":    rows,
+		}
+		w.Header().Set("Content-Type", "application/json")
+		_ = json.NewEncoder(w).Encode(body)
+	}))
+	defer srv.Close()
+
+	run := func(max int64) (int64, int64) {
+		s := &SmartsheetSource{client: httpclient.New(httpclient.WithBaseURL(srv.URL))}
+		results, err := s.read(context.Background(), "123", source.ReadOptions{MaxBatchBytes: max})
+		if err != nil {
+			t.Fatal(err)
+		}
+		var b, rw int64
+		for res := range results {
+			if res.Err != nil {
+				t.Fatal(res.Err)
+			}
+			b++
+			rw += res.Batch.NumRows()
+			res.Batch.Release()
+		}
+		return b, rw
+	}
+
+	offB, offR := run(0)
+	onB, onR := run(4096)
+	if offB != 1 {
+		t.Fatalf("cap-off batches=%d want 1", offB)
+	}
+	if onB <= 1 {
+		t.Fatalf("cap-on batches=%d want >1", onB)
+	}
+	if offR != onR || offR != 50 {
+		t.Fatalf("row mismatch off=%d on=%d", offR, onR)
 	}
 }
