+++ pkg/source/snapchatads/snapchatads_test.go
@@ -1,9 +1,14 @@
 package snapchatads
 
 import (
+	"context"
 	"encoding/json"
+	"net/http"
+	"net/http/httptest"
+	"strings"
 	"testing"
 
+	httpclient "github.com/bruin-data/ingestr/pkg/http"
 	"github.com/bruin-data/ingestr/pkg/source"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -481,4 +486,54 @@ func TestParseStatsTableWithDimensionAndPivot(t *testing.T) {
 	assert.Equal(t, "GEO", result.config.dimension)
 	assert.Equal(t, "country", result.config.pivot)
 	assert.Equal(t, "impressions", result.config.fields)
+}
+
+func TestSnapchatAdsByteCap(t *testing.T) {
+	wide := strings.Repeat("x", 2048)
+	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		wrappers := make([]map[string]interface{}, 0, 50)
+		for i := 0; i < 50; i++ {
+			wrappers = append(wrappers, map[string]interface{}{
+				"sub_request_status": "SUCCESS",
+				"organization":       map[string]interface{}{"id": i, "name": wide},
+			})
+		}
+		body := map[string]interface{}{
+			"request_status": "SUCCESS",
+			"organizations":  wrappers,
+		}
+		w.Header().Set("Content-Type", "application/json")
+		_ = json.NewEncoder(w).Encode(body)
+	}))
+	defer srv.Close()
+
+	run := func(max int64) (int64, int64) {
+		s := &SnapchatAdsSource{client: httpclient.New(httpclient.WithBaseURL(srv.URL))}
+		results, err := s.read(context.Background(), "organizations", nil, nil, source.ReadOptions{MaxBatchBytes: max})
+		if err != nil {
+			t.Fatal(err)
+		}
+		var b, rw int64
+		for res := range results {
+			if res.Err != nil {
+				t.Fatal(res.Err)
+			}
+			b++
+			rw += res.Batch.NumRows()
+			res.Batch.Release()
+		}
+		return b, rw
+	}
+
+	offB, offR := run(0)
+	onB, onR := run(4096)
+	if offB != 1 {
+		t.Fatalf("cap-off batches=%d want 1", offB)
+	}
+	if onB <= 1 {
+		t.Fatalf("cap-on batches=%d want >1", onB)
+	}
+	if offR != onR || offR != 50 {
+		t.Fatalf("row mismatch off=%d on=%d", offR, onR)
+	}
 }
