+++ pkg/source/trello/trello_test.go
@@ -1,8 +1,17 @@
 package trello
 
 import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"net/http/httptest"
+	"strings"
 	"testing"
 	"time"
+
+	httpclient "github.com/bruin-data/ingestr/pkg/http"
+	"github.com/bruin-data/ingestr/pkg/source"
 )
 
 func TestParseURI(t *testing.T) {
@@ -246,4 +255,47 @@ func idsOf(items []map[string]interface{}) []string {
 		}
 	}
 	return ids
+}
+
+func TestTrelloByteCap(t *testing.T) {
+	wide := strings.Repeat("x", 2048)
+	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		rows := make([]map[string]interface{}, 0, 50)
+		for i := 0; i < 50; i++ {
+			rows = append(rows, map[string]interface{}{"id": fmt.Sprintf("%d", i), "name": wide})
+		}
+		w.Header().Set("Content-Type", "application/json")
+		_ = json.NewEncoder(w).Encode(rows)
+	}))
+	defer srv.Close()
+
+	run := func(max int64) (int64, int64) {
+		s := &TrelloSource{client: httpclient.New(httpclient.WithBaseURL(srv.URL))}
+		results, err := s.read(context.Background(), "boards", source.ReadOptions{MaxBatchBytes: max})
+		if err != nil {
+			t.Fatal(err)
+		}
+		var b, rw int64
+		for res := range results {
+			if res.Err != nil {
+				t.Fatal(res.Err)
+			}
+			b++
+			rw += res.Batch.NumRows()
+			res.Batch.Release()
+		}
+		return b, rw
+	}
+
+	offB, offR := run(0)
+	onB, onR := run(4096)
+	if offB != 1 {
+		t.Fatalf("cap-off batches=%d want 1", offB)
+	}
+	if onB <= 1 {
+		t.Fatalf("cap-on batches=%d want >1", onB)
+	}
+	if offR != onR || offR != 50 {
+		t.Fatalf("row mismatch off=%d on=%d", offR, onR)
+	}
 }
